test: 컨트롤러 단위 테스트 클래스 분리 (포인트 조회, 포인트 충전)
