+++ src/test/java/io/hhplus/tdd/point/presentation/PointControllerChargeUnitTest.java
@@ -2,7 +2,6 @@ package io.hhplus.tdd.point.presentation;
 
 import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.Mockito.when;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.patch;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
@@ -23,8 +22,9 @@ import io.hhplus.tdd.point.exception.InvalidChargingPointException;
 import io.hhplus.tdd.point.exception.InvalidUserIdException;
 import io.hhplus.tdd.point.exception.UserNotFoundException;
 
+@DisplayName("포인트 충천 컨트롤러 단위 테스트")
 @WebMvcTest(PointController.class)
-class PointControllerTest {
+class PointControllerChargeUnitTest {
 
 	@Autowired
 	private MockMvc mockMvc;
@@ -32,88 +32,7 @@ class PointControllerTest {
 	@MockBean
 	private UserPointService userPointService;
 
-	@DisplayName("[포인트 조회 실패] 음수인 유저 식별자인 경우 bad request 를 반환한다")
-	@Test
-	void failBecauseOfUserIdNegative() throws Exception {
-		// given
-		long userId = -1L;
-
-		when(userPointService.findById(anyLong())).thenThrow(new InvalidUserIdException());
-
-		// when
-		final ResultActions result = mockMvc.perform(get("/point/{id}", userId)
-			.contentType(MediaType.APPLICATION_JSON));
-
-		// then
-		result.andExpectAll(
-			status().isBadRequest(),
-			jsonPath("$.code").value("MEMBER01"),
-			jsonPath("$.message").exists()
-		);
-	}
-
-	@DisplayName("[포인트 조회 실패] 0인 유저 식별자인 경우 bad request 를 반환한다")
-	@Test
-	void failBecauseOfUserIdZero() throws Exception {
-		// given
-		long userId = 0L;
-
-		when(userPointService.findById(anyLong())).thenThrow(new InvalidUserIdException());
-
-		// when
-		final ResultActions result = mockMvc.perform(get("/point/{id}", userId)
-			.contentType(MediaType.APPLICATION_JSON));
-
-		// then
-		result.andExpectAll(
-			status().isBadRequest(),
-			jsonPath("$.code").value("MEMBER01"),
-			jsonPath("$.message").exists()
-		);
-	}
-
-	@DisplayName("[포인트 조회 실패] 등록되지 않은 유저 식별자인 경우 bad request 를 반환한다")
-	@Test
-	void failBecauseOfInvalidUserId() throws Exception {
-		// given
-		long userId = 3L;
-
-		when(userPointService.findById(anyLong())).thenThrow(new UserNotFoundException());
-
-		// when
-		final ResultActions result = mockMvc.perform(get("/point/{id}", userId)
-			.contentType(MediaType.APPLICATION_JSON));
-
-		// then
-		result.andExpectAll(
-			status().isBadRequest(),
-			jsonPath("$.code").value("MEMBER02"),
-			jsonPath("$.message").exists()
-		);
-	}
-
-	@DisplayName("[포인트 조회 성공] 등록된 유저 식별자인 경우 ok 를 반환한다")
-	@Test
-	void successFindingUserPoint() throws Exception {
-		// given
-		long userId = 1L;
-		long point = 300L;
-
-		when(userPointService.findById(anyLong())).thenReturn(new UserPointSelectDTO(1L, 300L));
-
-		// when
-		final ResultActions result = mockMvc.perform(get("/point/{id}", userId)
-			.contentType(MediaType.APPLICATION_JSON));
-
-		// then
-		result.andExpectAll(
-			status().isOk(),
-			jsonPath("$.userId").value(userId),
-			jsonPath("$.point").value(point)
-		);
-	}
-
-	@DisplayName("[포인트 충전 실패] 충전 포인트가 음수이면 bad request 를 반환한다")
+	@DisplayName("[실패] 충전 포인트가 음수인 경우, bad request 를 반환한다")
 	@Test
 	void failChargingBecauseOfChargingPointNegative() throws Exception {
 		// given
@@ -136,7 +55,7 @@ class PointControllerTest {
 		);
 	}
 
-	@DisplayName("[포인트 충전 실패] 충전 포인트가 0이면 bad request 를 반환한다")
+	@DisplayName("[실패] 충전 포인트가 0이면, bad request 를 반환한다")
 	@Test
 	void failChargingBecauseOfChargingPointZero() throws Exception {
 		// given
@@ -159,7 +78,7 @@ class PointControllerTest {
 		);
 	}
 
-	@DisplayName("[포인트 충전 실패] 충전 포인트가 최대 충전 가능 포인트보다 크면 bad request 를 반환한다")
+	@DisplayName("[실패] 충전 포인트가 최대 충전 가능 포인트보다 큰 경우, bad request 를 반환한다")
 	@Test
 	void failChargingBecauseOfMaxChargingPoint() throws Exception {
 		// given
@@ -182,7 +101,7 @@ class PointControllerTest {
 		);
 	}
 
-	@DisplayName("[포인트 충전 실패] 유저 식별자가 음수인 경우 bad request 를 반환한다")
+	@DisplayName("[실패] 유저 식별자가 음수인 경우, bad request 를 반환한다")
 	@Test
 	void failChargingBecauseOfUserIdNegative() throws Exception {
 		// given
@@ -205,7 +124,7 @@ class PointControllerTest {
 		);
 	}
 
-	@DisplayName("[포인트 충전 실패] 유저 식별자가 0인 경우 bad request 를 반환한다")
+	@DisplayName("[실패] 유저 식별자가 0인 경우, bad request 를 반환한다")
 	@Test
 	void failChargingBecauseOfUserIdZero() throws Exception {
 		// given
@@ -228,7 +147,7 @@ class PointControllerTest {
 		);
 	}
 
-	@DisplayName("[포인트 충전 실패] 유저 식별자가 0인 경우 bad request 를 반환한다")
+	@DisplayName("[실패] 유저 식별자가 0인 경우, bad request 를 반환한다")
 	@Test
 	void failChargingBecauseOfUnRegisteredUserId() throws Exception {
 		// given
@@ -251,7 +170,7 @@ class PointControllerTest {
 		);
 	}
 
-	@DisplayName("[포인트 충전 성공] 등록된 유저 식별자와 단일 최대 충전 포인트를 넘지 않는 양수의 충전 포인트는 ok 를 반환한다")
+	@DisplayName("[성공] 등록된 유저 식별자와 단일 최대 충전 포인트를 넘지 않는 양수의 충전 포인트인 경우, ok 를 반환한다")
 	@Test
 	void successCharging() throws Exception {
 		// given
